feat(job-matcher): enhance job matching functionality and email integration
- Update job and match date fields from u64 to f32 for better precision
- Add GetJobs type to return both jobs and matching jobs
- Implement Mailjet email service with advanced features
- Improve job matching logic and UI integration
- Add error handling for permission checks
- Simplify landing page component structure
```

This commit message:
1. Uses "feat" type since these are new features/enhancements
2. Specifies "job-matcher" scope where appropriate
3. Provides a concise description under 50 chars in the subject line
4. Includes a body that lists the major functional changes without repeating the subject
5. Focuses on the significant changes rather than every file modification
6. Uses imperative mood and avoids periods
7. Groups related changes together logically

+++ src/backend/src/job_matcher/impliment.rs
@@ -5,7 +5,7 @@ use ic_cdk::caller;
 
 impl Job {
     pub fn new(id:String) -> Self {
-        let now = ic_cdk::api::time() as u64;
+        let now = ic_cdk::api::time() as f32;
 
         Job {
             notification_id: String::new(),
@@ -35,7 +35,7 @@ impl Job {
 
 
     pub fn update(&mut self, field_name: &str, data: Vec<String>) {
-        self.date_updated = ic_cdk::api::time() as u64;
+        self.date_updated = ic_cdk::api::time() as f32;
 
         match field_name {
             "skills" => self.skills = data,
@@ -140,7 +140,7 @@ impl Job {
                     Err("Permission denied".to_string())
                 },
                 Some(job) => {
-                    let now = ic_cdk::api::time() as u64;
+                    let now = ic_cdk::api::time() as f32;
                     
                     let mut updated_job = job.clone();
                     updated_job.skills = Vec::new();
@@ -206,7 +206,7 @@ impl Job {
     
             // Update the job's matches
             job.matches = matches;
-            job.date_updated = ic_cdk::api::time() as u64;
+            job.date_updated = ic_cdk::api::time() as f32;
             
             // Save the updated job
             store.insert(job_id.clone(), job.clone());
@@ -220,7 +220,7 @@ impl Job {
                         job_id: job_id.clone(),
                         user_id: job.user_id.clone(),
                         missmatching_skills: match_item.missmatching_skills.clone(),
-                        date_updated: ic_cdk::api::time() as u64,
+                        date_updated: ic_cdk::api::time() as f32,
                         is_connected: match_item.is_connected,
                     };
     
@@ -229,7 +229,7 @@ impl Job {
                     
                     // Add the new reciprocal match
                     other_job.matches.push(reciprocal_match);
-                    other_job.date_updated = ic_cdk::api::time() as u64;
+                    other_job.date_updated = ic_cdk::api::time() as f32;
                     
                     // Save the other job
                     store.insert(match_item.job_id.clone(), other_job);
